harden capability for fields

+++ includes/fields/class-listing.php
@@ -162,6 +162,10 @@ class Listing extends Field {
 	 */
 	public function create( $args = [] ) {
 
+		if ( ! current_user_can( 'manage_options' ) ) {
+			return;
+		}
+
 		if ( ! isset( $args['name'] ) || empty( $args['name'] ) ) {
 			throw new \InvalidArgumentException( sprintf( __( 'Can\'t find property %s' ), 'name' ) );
 		}
@@ -237,6 +241,10 @@ class Listing extends Field {
 	 * @return void
 	 */
 	public function delete() {
+
+		if ( ! current_user_can( 'manage_options' ) ) {
+			return;
+		}
 
 		wp_delete_post( $this->get_post_id(), true );
